Revert "feature(testing): Use candidate versions of pepr in E2E tests"
