+++ _helpers/package.json
@@ -11,7 +11,6 @@
     "@jest/globals": "^29.7.0",
     "@types/node": "^20.13.0",
     "commander": "^11.1.0",
-    "find-up": "^7.0.0",
     "jest": "^29.7.0",
     "kubernetes-fluent-client": "^2.6.1",
     "ts-jest": "^29.1.4",
