--- test-specific-version/README.md
@@ -1,3 +1,0 @@
-# Use Arbitrary Pepr Version
-
-This example shows how to run tests across multiple versions of pepr, to include release candidates on a branch not yet merged into `main`.
