--- test-specific-version/jest.config.json
@@ -1,4 +1,0 @@
-{
-  "preset": "ts-jest",
-  "testEnvironment": "node"
-}
