--- test-specific-version/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "test-specific-version",
-  "version": "0.0.1",
-  "description": "Pepr/#1109",
-  "keywords": [
-    "pepr",
-    "k8s",
-    "pepr-module",
-    "testing"
-  ],
-  "engines": {
-    "node": ">=18.0.0"
-  },
-  "pepr": {
-    "uuid": "6b7cfad2-722b-47f8-99ed-f3cbf65ab5b1",
-    "onError": "reject",
-    "webhookTimeout": 10,
-    "customLabels": {
-      "namespace": {
-        "pepr.dev": ""
-      }
-    },
-    "alwaysIgnore": {
-      "namespaces": []
-    },
-    "includedFiles": [],
-    "env": {
-      "LOG_LEVEL": "debug"
-    }
-  },
-  "dependencies": {
-    "pepr": "^0.31.1"
-  },
-  "devDependencies": {
-    "@jest/globals": "^29.7.0",
-    "@types/node": "^20.13.0",
-    "find-up": "^5.0.0",
-    "jest": "^29.7.0",
-    "kubernetes-fluent-client": "^2.6.1",
-    "typescript": "5.3.3"
-  },
-  "scripts": {
-    "pepr": "pepr",
-    "cli": "npm run --workspace helpers cli -- --module=$(pwd)",
-    "test": "npm run cli -- test --suite all",
-    "test:e2e": "npm run cli -- test --suite e2e",
-    "test:unit": "npm run cli -- test --suite unit"
-  }
-}
