--- test-specific-version/tsconfig.json
@@ -1,28 +1,0 @@
-{
-  "compilerOptions": {
-    "allowSyntheticDefaultImports": true,
-    "declaration": true,
-    "declarationMap": true,
-    "emitDeclarationOnly": true,
-    "esModuleInterop": true,
-    "lib": [
-      "ES2022"
-    ],
-    "module": "CommonJS",
-    "moduleResolution": "node",
-    "outDir": "dist",
-    "resolveJsonModule": true,
-    "rootDir": ".",
-    "strict": false,
-    "target": "ES2022",
-    "useUnknownInCatchVariables": false
-  },
-  "include": [
-    "**/*.ts"
-  ],
-  "exclude": [
-    "**/*.test.*",
-    "./dist/**/*",
-    "./dev/**/*"
-  ]
-}
